Move database fixtures to conftest for general use
